+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 - enviroment only accept correct types
 - can detect physical/magical damage type to create the damage action
 - can detect direct damage
+- It can detect spend hope to create actions
 
 # 0.1.5
 - import enviroment/adversary can also import the features the world